requirements.txt
asgiref==3.9.2
bidict==0.23.1
blinker==1.9.0
brotli==1.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.1
colorama==0.4.6
ConfigArgParse==1.7.1
Django==5.2.7
django-cors-headers==4.9.0
django-environ==0.12.0
django-filter==25.1
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
Flask==3.1.2
flask-cors==6.0.2
Flask-Login==0.6.3
gevent==25.9.1
geventhttpclient==2.3.7
greenlet==3.3.0
h11==0.16.0
idna==3.10
inflection==0.5.1
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
locust==2.43.0
MarkupSafe==3.0.3
msgpack==1.1.2
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
psutil==7.2.1
psycopg2-binary==2.9.10
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.2
python-engineio==4.13.0
python-socketio==5.16.0
pytz==2025.2
pywin32==311
PyYAML==6.0.3
pyzmq==27.1.0
redis==7.1.0
requests==2.32.5
simple-websocket==1.1.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
websocket-client==1.9.0
Werkzeug==3.1.4
wsproto==1.3.2
zope.event==6.1
zope.interface==8.1.1
